+++ src/lib/dhcp/tests/option_data_types_unittest.cc
@@ -59,7 +59,7 @@ public:
             break;
         default:
             // This loop is incorrectly compiled by some old g++?!
-            for (int i = 0; i < sizeof(T); ++i) {
+            for (size_t i = 0; i < sizeof(T); ++i) {
                 buf.push_back(value >> ((sizeof(T) - i - 1) * 8) & 0xFF);
             }
         }
